Extract system_payment module to a file.
