+++ src/test_data.rs
@@ -10,6 +10,7 @@ use rand::{prelude::*, Rng};
 use crate::sample::Sample;
 
 mod native_transfer;
+mod system_payment;
 
 // From the chainspec.
 // 1 minute.
@@ -99,35 +100,6 @@ impl TransferTarget {
             TransferTarget::URef(_) => "target:uref",
             TransferTarget::Key(_) => "target:key-account",
         }
-    }
-}
-
-mod system_payment {
-    use casper_execution_engine::core::engine_state::ExecutableDeployItem;
-    use casper_types::{bytesrepr::Bytes, runtime_args, RuntimeArgs, U512};
-
-    use crate::sample::Sample;
-
-    pub(super) fn valid() -> Sample<ExecutableDeployItem> {
-        let payment = ExecutableDeployItem::ModuleBytes {
-            module_bytes: Bytes::new(),
-            args: runtime_args! {
-                "amount" => U512::from(1000000000)
-            },
-        };
-
-        Sample::new("payment:system", payment, true)
-    }
-
-    pub(super) fn invalid() -> Sample<ExecutableDeployItem> {
-        let payment = ExecutableDeployItem::ModuleBytes {
-            module_bytes: Bytes::new(),
-            args: runtime_args! {
-                "paying" => U512::from(1000000000)
-            },
-        };
-
-        Sample::new("payment:system-missing:amount", payment, false)
     }
 }
